// src/data/vrboPulled.generated.ts
/* eslint-disable */
/**
 * AUTO-GENERATED by scripts/pull-vrbo.ts
 * Edit src/data/units.ts if you want to override copy.
 */
export const vrboPulled = [
  {
    "unitId": "35",
    "vrboUrl": "https://www.vrbo.com/en-ca/cottage-rental/p2160974vb",
    "images": [],
    "amenities": [
      "tv"
    ]
  },
  {
    "unitId": "87",
    "vrboUrl": "https://www.vrbo.com/en-ca/cottage-rental/p3511146vb",
    "images": [],
    "amenities": [
      "tv"
    ]
  },
  {
    "unitId": "128",
    "vrboUrl": "https://www.vrbo.com/en-ca/cottage-rental/p1300156vb",
    "images": [],
    "amenities": [
      "tv"
    ]
  },
  {
    "unitId": "151",
    "vrboUrl": "https://www.vrbo.com/en-ca/cottage-rental/p1587209vb",
    "images": [],
    "amenities": [
      "tv"
    ]
  },
  {
    "unitId": "56",
    "vrboUrl": "https://www.vrbo.com/en-ca/cottage-rental/p3925710vb",
    "images": [],
    "amenities": [
      "tv"
    ]
  }
] as const;